telegram-bot-backend/pyproject.toml
[project]
name = "telegram-bot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.0.0,<4.0.0",
    "aiohttp>=3.11.18",
    "python-dotenv>=1.1.0",
]
